New Docker

+++ docker/build.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
 #
 # build.sh: A simple script to build the Docker image.  It is intended to be run
-# here under the docker directoxpry, in a local git clone repo where the pip
-# package has already been built.
+# here under the docker directory, in a local git clone repo.
 #
 # (c) 2023/24 RTE
 # Developed by Grupo AIA
@@ -16,21 +15,22 @@ set -o nounset -o noclobber
 set -o errexit -o pipefail 
 
 
-# Ask for the container TAG
-if [[ $# -ne 1 ]]; then
+# Ask for the container TAG and Dynawo path
+if [[ $# -ne 2 ]]; then
     echo
-    echo -e "$0: A tag is required for the image to be built.\n"
+    echo -e "$0: A tag and the path to the Dynawo installation are required.\n"
+    echo -e "Usage: $0 <TAG> <DYNAWO_HOST_PATH>\n"
     echo
     exit 4
 fi
 TAG=$1
+DYNAWO_HOST_PATH=$2
 
 # The pip package needs to appear at the build directory level
 if [ ! -d ../dist ]; then
   cd ..
-  pip install build
-  echo "Building package..."
-  python3 -m build
+  echo "Building package with uv..."
+  uv build --out-dir dist
   cd docker
   PKG=$(find ../dist -iname '*.whl' -printf "%Ts %P\n" | sort -n | tail -n 1 | cut -d' ' -f2)  # newest wheel
   rm -f "$PKG"
@@ -46,19 +46,29 @@ fi
 EXAMPLES=examples
 cp -a ../examples .
 
-# Dynawo: instead of downloading, unpacking, and using COPY, we'll do it all in
-# the Dockerfile at build time.  (Note: if the Dynawo ZIP was published as a
-# tar.gz, we could download it here and use ADD instead.)
-ZIP="Dynawo_omc_V1.6.0.zip"
+# Dynawo: Copy the host directory to the build context
+DYNAWO_DIR_NAME=dynawo_build
+if [ ! -d "$DYNAWO_HOST_PATH" ]; then
+   echo "ERROR: Dynawo path $DYNAWO_HOST_PATH not found."
+   exit 1
+fi
+rm -rf "$DYNAWO_DIR_NAME" # Clean up previous
+cp -a "$DYNAWO_HOST_PATH" "$DYNAWO_DIR_NAME"
+
+# Check if the expected executable exists
+if [ ! -f "$DYNAWO_DIR_NAME/dynawo/dynawo.sh" ]; then
+    echo "ERROR: Expected executable not found at '$DYNAWO_HOST_PATH/dynawo/dynawo.sh'"
+    rm -rf "$DYNAWO_DIR_NAME" # Clean up
+    exit 1
+fi
 
 # Launch the build
 rm -f build.log
 docker build -t dycov:latest -t dycov:"$TAG" \
              --build-arg dycov_PKG="$PKG" \
              --build-arg dycov_EXAMPLES="$EXAMPLES" \
-             --build-arg DWO_ZIP="$ZIP" \
+             --build-arg DYNAWO_DIR_NAME="$DYNAWO_DIR_NAME" \
              .
 
 # Clean up
-rm -rf "$PKG" "$EXAMPLES"
-
+rm -rf "$PKG" "$EXAMPLES" "$DYNAWO_DIR_NAME"
